add javadoc to a larger part of the current codebase. Clean up some client sided code that wont be used anymore.
