--- src/main/java/inf112/skeleton/server/util/ResourceLoader.java
@@ -1,11 +1,0 @@
-package inf112.skeleton.server.util;
-
-import com.badlogic.gdx.maps.tiled.TmxMapLoader;
-
-public class ResourceLoader {
-    public static String loadResource(){
-        ClassLoader classLoader = ResourceLoader.class.getClassLoader();
-        return classLoader.getResource("Board.tmx").getPath();
-
-    }
-}
